Move freeze functions into utility
